добавляет исправления в модуль отправки данных и вывод popup
