+++ js/main.js
@@ -1,5 +1,4 @@
-import { setFormListeners } from './validation.js';
+import './validation.js';
 import './map.js';
 import './server.js';
 
-setFormListeners();
