+++ js/popup.js
@@ -1,8 +1,11 @@
-const successTemplate = document.querySelector('#success').content;
-const errorTemplate = document.querySelector('#error').content;
+import { resetSettings } from './map.js';
+
+const successTemplate = document.querySelector('#success').content.querySelector('.success');
+const errorTemplate = document.querySelector('#error').content.querySelector('.error');
 const errorButton = errorTemplate.querySelector('.error__button');
 
 const ESC_KEY = 27;
+let closeMessage = null;
 
 const onDocumentClick = () => {
   closeMessage();
@@ -10,13 +13,14 @@ const onDocumentClick = () => {
 
 const onDocumentKeydown = (evt) => {
   if (evt.key === ESC_KEY) {
-    closeMessage(); // не знаю как объявить функцию, чтобы принимала данные находящиеся ниже нее...
+    closeMessage();
   }
 };
 
 const showSuccessMessage = () => {
   const successMessage = successTemplate.cloneNode(true);
   document.body.appendChild(successMessage);
+  resetSettings();
   document.addEventListener('click', onDocumentClick);
   errorButton.addEventListener('click', onDocumentClick);
   document.addEventListener('keydown', onDocumentKeydown);
@@ -34,7 +38,7 @@ const showErrorMessage = (message) => {
   document.addEventListener('keydown', onDocumentKeydown);
 };
 
-const closeMessage = function () {
+closeMessage = () => {
   const successMessagePopup = document.querySelector('.success');
   const errorMessagePopup = document.querySelector('.error');
