+++ js/util.js
@@ -1,40 +1,4 @@
-const getRandomPositiveFloat = (a, b, digits = 1) => {
-  const lower = Math.min(Math.abs(a), Math.abs(b));
-  const upper = Math.max(Math.abs(a), Math.abs(b));
-  const result = Math.random() * (upper - lower) + lower;
-
-  return Number(result.toFixed(digits));
-};
-
-const getRandomPositiveInteger = (a, b) => {
-  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
-  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
-  const result = Math.random() * (upper - lower + 1) + lower;
-
-  return Math.floor(result);
-};
-
-const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length -1)];
-
-const shuffleArray = (array) => {
-  const mixedArray = array.slice();
-
-  for (let i = mixedArray.length - 1; i > 0; i--) {
-    const randomIndex = Math.floor(Math.random() * (i + 1));
-    const tempValue = mixedArray[i];
-    mixedArray[i] = mixedArray[randomIndex];
-    mixedArray[randomIndex] = tempValue;
-  }
-
-  return mixedArray;
-};
-
-const getRandomArray = (arr) => {
-  const shuffledArray = shuffleArray(arr);
-  const random = getRandomPositiveInteger(0, arr.length -1);
-
-  return shuffledArray.slice(0, random);
-};
+const ALERT_SHOW_TIME = 5000;
 
 const showAlert = (message) => {
   const alertContainer = document.createElement('div');
@@ -54,8 +18,8 @@ const showAlert = (message) => {
 
   setTimeout(() => {
     alertContainer.remove();
-  }, 5000);
+  }, ALERT_SHOW_TIME);
 };
 
 
-export {getRandomPositiveFloat, getRandomPositiveInteger, getRandomArrayElement, getRandomArray, showAlert};
+export { showAlert };
